Safari-611.1.10/LayoutTests/media/track/video/video-track-mkv-theora-addtrack-expected.txt


Testing video track "addtrack" events.

EVENT(addtrack)
EXPECTED (video.videoTracks contains 'event.track') OK

EVENT(addtrack)
EXPECTED (video.videoTracks contains 'event.track') OK

EVENT(canplaythrough)
END OF TEST
